+++ src/disable__ontouchmove/index.d.ts
@@ -1,5 +1,0 @@
-export declare function disable__ontouchmove(e:TouchEvent):void
-export {
-	disable__ontouchmove as ontouchmove_disable,
-	disable__ontouchmove as ontouchmove__disable,
-}