+++ src/enable__ontouchmove/index.d.ts
@@ -1,5 +1,0 @@
-export declare function enable__ontouchmove():boolean
-export {
-	enable__ontouchmove as ontouchmove_enable,
-	enable__ontouchmove as ontouchmove__enable,
-}