+++ src/enable__ontouchmove/index.js
@@ -1,7 +1,0 @@
-export function enable__ontouchmove() {
-	return true
-}
-export {
-	enable__ontouchmove as ontouchmove_enable,
-	enable__ontouchmove as ontouchmove__enable,
-}