+++ src/index.js
@@ -1,5 +1,5 @@
-export * from './disable__ontouchmove/index.js'
-export * from './enable__ontouchmove/index.js'
+export * from './ontouchmove__disable/index.js'
+export * from './ontouchmove__enable/index.js'
 export * from './no_touch__mount/index.js'
 export * from './touch_scrolling__disable/index.js'
 export * from './touch_scrolling__enable/index.js'