+++ src/ontouchmove__disable/index.d.ts
@@ -1,0 +1,5 @@
+export declare function ontouchmove__disable(e:TouchEvent):void
+export {
+	ontouchmove__disable as disable__ontouchmove,
+	ontouchmove__disable as ontouchmove_disable,
+}