+++ src/ontouchmove__disable/index.js
@@ -1,0 +1,10 @@
+/**
+ * @param e{TouchEvent}
+ */
+export function ontouchmove__disable(e) {
+	e.preventDefault()
+}
+export {
+	ontouchmove__disable as disable__ontouchmove,
+	ontouchmove__disable as ontouchmove_disable,
+}