+++ src/ontouchmove__enable/index.d.ts
@@ -1,0 +1,5 @@
+export declare function ontouchmove__enable():boolean
+export {
+	ontouchmove__enable as enable__ontouchmove,
+	ontouchmove__enable as ontouchmove_enable,
+}