+++ src/ontouchmove__enable/index.js
@@ -1,0 +1,7 @@
+export function ontouchmove__enable() {
+	return true
+}
+export {
+	ontouchmove__enable as enable__ontouchmove,
+	ontouchmove__enable as ontouchmove_enable,
+}